Put sitemap classes under a module
