+++ lib/rummager.rb
@@ -176,8 +176,6 @@ require "search_config"
 require "search_server"
 require "services"
 require "sitemap/property_boost_calculator"
-require "sitemap/sitemap"
-require "sitemap/sitemap_cleanup"
-require "sitemap/sitemap_generator"
-require "sitemap/sitemap_presenter"
-require "sitemap/sitemap_writer"
+require "sitemap/generator"
+require "sitemap/presenter"
+require "sitemap/writer"
