+++ spec/unit/sitemap/generator_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-RSpec.describe SitemapGenerator do
+RSpec.describe Sitemap::Generator do
   before do
     allow_any_instance_of(LegacyClient::IndexForSearch).to receive(:real_index_names).and_return(%w(govuk_test))
   end
