+++ spec/unit/sitemap/sitemap_presenter_spec.rb
@@ -1,5 +1,4 @@
 require "spec_helper"
-require "sitemap/sitemap"
 
 RSpec.describe SitemapPresenter do
   before do
